优化数字的出现

+++ app/src/main/java/com/zgmao/game/bean/PlayingCard.java
@@ -13,10 +13,10 @@ package com.zgmao.game.bean;
 public class PlayingCard
 {
     private String name;
-    private double x;
-    private double y;
-    private double width;
-    private double height;
+    private int x;
+    private int y;
+    private int width;
+    private int height;
     private int number;
 
     public String getName()
@@ -29,42 +29,42 @@ public class PlayingCard
         this.name = name;
     }
 
-    public double getX()
+    public int getX()
     {
         return x;
     }
 
-    public void setX(double x)
+    public void setX(int x)
     {
         this.x = x;
     }
 
-    public double getY()
+    public int getY()
     {
         return y;
     }
 
-    public void setY(double y)
+    public void setY(int y)
     {
         this.y = y;
     }
 
-    public double getWidth()
+    public int getWidth()
     {
         return width;
     }
 
-    public void setWidth(double width)
+    public void setWidth(int width)
     {
         this.width = width;
     }
 
-    public double getHeight()
+    public int getHeight()
     {
         return height;
     }
 
-    public void setHeight(double height)
+    public void setHeight(int height)
     {
         this.height = height;
     }
